Fix tests, support binary and keyword fields with ignore above for pluggable data formats
