+++ sandbox/plugins/parquet-data-format/src/main/java/org/opensearch/parquet/writer/ParquetDocumentInput.java
@@ -22,6 +22,7 @@ import org.opensearch.parquet.ParquetDataFormatPlugin;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Stream;
 
 /**
  * Document input for the Parquet data format.
@@ -47,6 +48,9 @@ public class ParquetDocumentInput implements DocumentInput<List<FieldValuePair>>
             .getOrDefault(ParquetDataFormatPlugin.PARQUET_DATA_FORMAT, Set.of());
         if (capabilities.isEmpty() && fieldType != PrimaryTermFieldType.INSTANCE) {
             // nothing to support on this format for this field.
+            if (Stream.of(Thread.currentThread().getStackTrace()).noneMatch(s -> s.getMethodName().contains("parseDynamicValue"))) {
+                logger.info("get none caps in non dynamic path: {} {}", fieldType.name(), fieldType.getCapabilityMap());
+            }
             logger.trace("Ignored to add field: {} {}", fieldType.name(), fieldType.getCapabilityMap());
             return;
         }
