+++ server/src/main/java/org/opensearch/index/engine/dataformat/FieldCapabilityAssigner.java
@@ -82,7 +82,10 @@ public class FieldCapabilityAssigner {
                     + configuredFormats.stream().map(DataFormat::name).collect(Collectors.toList())
             );
         }
-        logger.debug("{} assigned to :  {}", fieldType, assigned);
+        logger.info("{} assigned to :  {}", fieldType.name(), assigned);
+        if (fieldType.name().equals("fieldA_10")) {
+            logger.info("{} assigned to :  {} for debug", fieldType.name(), assigned);
+        }
         fieldType.setCapabilityMap(Map.copyOf(assigned));
     }
