+++ server/src/main/java/org/opensearch/index/mapper/FilterFieldType.java
@@ -20,6 +20,8 @@ import org.opensearch.common.geo.ShapeRelation;
 import org.opensearch.common.time.DateMathParser;
 import org.opensearch.common.unit.Fuzziness;
 import org.opensearch.index.analysis.NamedAnalyzer;
+import org.opensearch.index.engine.dataformat.DataFormat;
+import org.opensearch.index.engine.dataformat.FieldTypeCapabilities;
 import org.opensearch.index.fielddata.IndexFieldData;
 import org.opensearch.index.query.IntervalMode;
 import org.opensearch.index.query.QueryRewriteContext;
@@ -31,6 +33,7 @@ import java.io.IOException;
 import java.time.ZoneId;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
@@ -289,5 +292,25 @@ public abstract class FilterFieldType extends MappedFieldType {
     @Override
     public MappedFieldType unwrap() {
         return delegate.unwrap();
+    }
+
+    @Override
+    public Map<DataFormat, Set<FieldTypeCapabilities.Capability>> getCapabilityMap() {
+        return delegate.getCapabilityMap();
+    }
+
+    @Override
+    public synchronized void setCapabilityMap(Map<DataFormat, Set<FieldTypeCapabilities.Capability>> capabilityMap) {
+        delegate.setCapabilityMap(capabilityMap);
+    }
+
+    @Override
+    protected FieldTypeCapabilities.Capability searchCapability() {
+        return delegate.searchCapability();
+    }
+
+    @Override
+    public Set<FieldTypeCapabilities.Capability> requestedCapabilities() {
+        return delegate.requestedCapabilities();
     }
 }
